Store create project form state in component
