+++ src/client/reducers.ts
@@ -1,9 +1,5 @@
 import { combineReducers } from "redux-loop";
 import authReducer, { AuthState, initialState as initialAuthState } from "./reducers/auth";
-import projectFormReducer, {
-  initialState as initialProjectFormState,
-  ProjectFormState
-} from "./reducers/projectForm";
 import projectsReducer, {
   initialState as initialProjectsState,
   ProjectsState
@@ -17,7 +13,6 @@ import userReducer, { initialState as initialUserState, UserState } from "./redu
 export interface State {
   readonly auth: AuthState;
   readonly user: UserState;
-  readonly projectForm: ProjectFormState;
   readonly regionConfig: RegionConfigState;
   readonly projects: ProjectsState;
 }
@@ -25,7 +20,6 @@ export interface State {
 export const initialState: State = {
   auth: initialAuthState,
   user: initialUserState,
-  projectForm: initialProjectFormState,
   regionConfig: initialRegionConfigState,
   projects: initialProjectsState
 };
@@ -33,7 +27,6 @@ export const initialState: State = {
 export default combineReducers({
   auth: authReducer,
   user: userReducer,
-  projectForm: projectFormReducer,
   regionConfig: regionConfigReducer,
   projects: projectsReducer
 });
